pyproject.toml
[project]
name = "make-a-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "dspy-ai>=3.0.3",
    "openai>=2.7.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.51.0",
    "uv>=0.9.19",
    "sync>=1.0.0",
]
